+++ app/services/user_service.py
@@ -1,8 +1,11 @@
 from sqlalchemy.ext.asyncio import AsyncSession
+
+from app.core.security import create_url_safe_token
 from .. import repositories as repo
+from ..models import User as UserModel
 from ..schemas.user import UserCreate, UserUpdate
 from ..utils.hashing import Hash
-from ..core.exceptions import AlreadyExistsException
+from ..core.exceptions import AlreadyExistsException, FlowerAppException
 
 
 async def register_new_user(db: AsyncSession, request: UserCreate):
@@ -19,12 +22,29 @@ async def register_new_user(db: AsyncSession, request: UserCreate):
 
         await db.flush()
         await db.refresh(new_user)
-    return new_user
+
+        #email
+        token = create_url_safe_token(
+            {"email" : new_user.email}
+        )
+    return new_user, token
+
 
 
 async def get_user_me(db: AsyncSession, user_id: int):
-    # Тут просто отримання, транзакція не обов'язкова, але можна через repo
     return await repo.user.get_user(db, user_id)
+
+
+
+async def get_all_users(
+    db: AsyncSession,
+    current_user: UserModel,
+    skip: int = 0,
+    limit: int = 100
+):
+    if current_user.role != "admin":
+        raise FlowerAppException("Тільки адміністратор може переглядати список користувачів")
+    return await repo.user.get_all(db, skip=skip, limit=limit)
 
 
 async def update_user_info(db: AsyncSession, user_id: int, user_update: UserUpdate):
@@ -50,7 +70,8 @@ async def update_user_info(db: AsyncSession, user_id: int, user_update: UserUpda
 
 
 async def delete_user_account(db: AsyncSession, user_id: int):
-    async with db.begin():
-        db_user = await repo.user.get_user(db, user_id)
+    db_user = await repo.user.get_user(db, user_id)
+    if db_user:
         await repo.user.delete_user(db, db_user)
+        await db.commit()
     return None
